practice/backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Configuration
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
authlib>=1.3.0
itsdangerous>=2.1.0

# LLM - OpenRouter/Qwen (uses OpenAI SDK)
openai>=1.12.0

# Qdrant Vector Database (lightweight client without fastembed)
qdrant-client>=1.7.1

# PostgreSQL (Neon) - Optional
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Utilities
pyyaml>=6.0.1
httpx>=0.26.0

# Development (not needed for production)
# pytest>=7.4.4
# pytest-asyncio>=0.23.3
